+++ Changelog.md
@@ -4,6 +4,7 @@
 - Fixes font glyph render in some corner cases where the rendered glyph did result in rectangle garbage on the screen.
 - Adds VT sequence for enabling/disabling debug logging. `CSI ? 46 h` and `CSI ? 46 l` and CLI option `-d`.
 - Adds improved debug logging.
+- Adds support for different font render modes: `lcd`, `light`, `gray`, `monochrome` in `profiles.NAME.font.render_mode` (default: `lcd`).
 - ...
 
 ### 0.1.1 (2020-12-31)